requirements.txt
langchain
pypdf
langchain_community
langchain-chroma
openai
langchain-openai
python-dotenv
matplotlib  